adjust backward facing step example to new precursor implementation
