+++ applications/incompressible_navier_stokes/backward_facing_step/application.h
@@ -19,323 +19,371 @@
  *  ______________________________________________________________________
  */
 
-#ifndef APPLICATIONS_INCOMPRESSIBLE_NAVIER_STOKES_TEST_CASES_TURBULENT_CHANNEL_H_
-#define APPLICATIONS_INCOMPRESSIBLE_NAVIER_STOKES_TEST_CASES_TURBULENT_CHANNEL_H_
+#ifndef APPLICATIONS_INCOMPRESSIBLE_NAVIER_STOKES_TEST_CASES_BFS_H_
+#define APPLICATIONS_INCOMPRESSIBLE_NAVIER_STOKES_TEST_CASES_BFS_H_
 
 // ExaDG
 #include <exadg/functions_and_boundary_conditions/linear_interpolation.h>
 
 // backward facing step application
+#include "include/functions.h"
 #include "include/geometry.h"
-#include "include/inflow_data_storage.h"
 #include "include/postprocessor.h"
 
 namespace ExaDG
 {
 namespace IncNS
 {
-template<int dim>
-class InitialSolutionVelocity : public dealii::Function<dim>
+// consider a friction Reynolds number of Re_tau = u_tau * H / nu = 290
+// and body force f = tau_w/H with tau_w = u_tau^2.
+double const viscosity = 1.5268e-5;
+
+// number of points for inflow boundary condition
+unsigned int const n_points_inflow = 101;
+
+// start and end time
+double const Re_H                 = 5540.0;
+double const centerline_velocity  = Re_H * viscosity / Geometry::H;
+double const characteristic_time  = Geometry::H / centerline_velocity;
+double const start_time           = 0.0;
+double const precursor_start_time = -300.0 * characteristic_time;
+double const end_time             = 300.0 * characteristic_time;
+
+// postprocessing
+
+// sampling of statistical results
+double const       sample_start_time      = 100.0 * characteristic_time;
+unsigned int const sample_every_timesteps = 10;
+unsigned int const n_points_per_line      = 101;
+
+// solver tolerances
+double const ABS_TOL = 1.e-12;
+double const REL_TOL = 1.e-3;
+
+double const ABS_TOL_LINEAR = 1.e-12;
+double const REL_TOL_LINEAR = 1.e-2;
+
+/*
+ *  Most of the parameters are the same for both domains, so we write
+ *  this function for the actual domain and only "correct" the parameters
+ *  for the precursor by passing an additional parameter is_precursor.
+ */
+void
+do_set_parameters(Parameters & param, bool const is_precursor = false)
 {
-public:
-  InitialSolutionVelocity(double const max_velocity,
-                          double const length,
-                          double const height,
-                          double const width)
-    : dealii::Function<dim>(dim, 0.0),
-      max_velocity(max_velocity),
-      length(length),
-      height(height),
-      width(width)
-  {
-  }
+  // MATHEMATICAL MODEL
+  param.problem_type                   = ProblemType::Unsteady;
+  param.equation_type                  = EquationType::NavierStokes;
+  param.use_outflow_bc_convective_term = true;
+  param.formulation_viscous_term       = FormulationViscousTerm::LaplaceFormulation;
+  param.formulation_convective_term    = FormulationConvectiveTerm::DivergenceFormulation;
+  param.right_hand_side                = true;
 
-  double
-  value(dealii::Point<dim> const & p, unsigned int const component = 0) const final
-  {
-    AssertThrow(dim == 3, dealii::ExcMessage("Dimension has to be dim==3."));
 
-    double const x = p[0] / length;
-    double const y = (p[1] - height / 2.0) / (height / 2.0);
-    double const z = p[2] / width;
+  // PHYSICAL QUANTITIES
+  if(is_precursor)
+    param.start_time = precursor_start_time;
+  else
+    param.start_time = start_time;
+  param.end_time  = end_time;
+  param.viscosity = viscosity;
 
-    double const factor = 0.5;
 
-    double result = 0.0;
-    if(false) // only random perturbations
-    {
-      if(component == 0)
-      {
-        if(std::abs(y) < 1.0)
-          result = -max_velocity * (pow(y, 6.0) - 1.0) *
-                   (1.0 + ((double)rand() / RAND_MAX - 0.5) * factor);
-        else
-          result = 0.0;
-      }
-    }
-    else // random perturbations and vortices
-    {
-      if(component == 0)
-      {
-        if(std::abs(y) < 1.0)
-          result = -max_velocity * (pow(y, 6.0) - 1.0) *
-                   (1.0 + (((double)rand() / RAND_MAX - 1.0) + std::sin(z * 8.0) * 0.5) * factor);
-      }
+  // TEMPORAL DISCRETIZATION
+  param.solver_type                     = SolverType::Unsteady;
+  param.temporal_discretization         = TemporalDiscretization::BDFDualSplittingScheme;
+  param.treatment_of_convective_term    = TreatmentOfConvectiveTerm::Explicit;
+  param.calculation_of_time_step_size   = TimeStepCalculation::CFL;
+  param.order_time_integrator           = 2;
+  param.start_with_low_order            = true;
+  param.adaptive_time_stepping          = true;
+  param.max_velocity                    = centerline_velocity;
+  param.cfl                             = 0.3;
+  param.cfl_exponent_fe_degree_velocity = 1.5;
+  param.time_step_size                  = 1.0e-1;
+  param.order_time_integrator           = 2;
 
-      if(component == 2)
-      {
-        if(std::abs(y) < 1.0)
-          result = -max_velocity * (pow(y, 6.0) - 1.0) * std::sin(x * 8.0) * 0.5 * factor;
-      }
-    }
+  // output of solver information
+  param.solver_info_data.interval_time = (end_time - start_time) / 100;
 
-    return result;
-  }
+  // SPATIAL DISCRETIZATION
+  param.grid.triangulation_type = TriangulationType::Distributed;
+  param.mapping_degree          = param.degree_u;
+  param.degree_p                = DegreePressure::MixedOrder;
 
-private:
-  double const max_velocity, length, height, width;
-};
+  // convective term
+  if(param.formulation_convective_term == FormulationConvectiveTerm::DivergenceFormulation)
+    param.upwind_factor = 0.5; // allows using larger CFL values for explicit formulations
 
-template<int dim>
-class InflowProfile : public dealii::Function<dim>
+  // variant Direct allows to use larger time step
+  // sizes due to CFL condition at inflow boundary
+  param.type_dirichlet_bc_convective = TypeDirichletBCs::Mirror;
+
+  // viscous term
+  param.IP_formulation_viscous = InteriorPenaltyFormulation::SIPG;
+
+  // velocity pressure coupling terms
+  param.gradp_formulation = FormulationPressureGradientTerm::Weak;
+  param.divu_formulation  = FormulationVelocityDivergenceTerm::Weak;
+
+  // div-div and continuity penalty
+  param.use_divergence_penalty                     = true;
+  param.divergence_penalty_factor                  = 1.0e0;
+  param.use_continuity_penalty                     = true;
+  param.continuity_penalty_factor                  = param.divergence_penalty_factor;
+  param.continuity_penalty_components              = ContinuityPenaltyComponents::Normal;
+  param.apply_penalty_terms_in_postprocessing_step = true;
+  param.continuity_penalty_use_boundary_data       = true;
+
+  // TURBULENCE
+  param.turbulence_model_data.is_active        = false;
+  param.turbulence_model_data.turbulence_model = TurbulenceEddyViscosityModel::Sigma;
+  // Smagorinsky: 0.165
+  // Vreman: 0.28
+  // WALE: 0.50
+  // Sigma: 1.35
+  param.turbulence_model_data.constant = 1.35;
+
+  // PROJECTION METHODS
+
+  // pressure Poisson equation
+  param.solver_pressure_poisson              = SolverPressurePoisson::CG;
+  param.solver_data_pressure_poisson         = SolverData(1e4, 1.e-12, 1.e-6, 100);
+  param.preconditioner_pressure_poisson      = PreconditionerPressurePoisson::Multigrid;
+  param.multigrid_data_pressure_poisson.type = MultigridType::cphMG;
+  param.multigrid_data_pressure_poisson.smoother_data.smoother = MultigridSmoother::Chebyshev;
+  param.multigrid_data_pressure_poisson.smoother_data.preconditioner =
+    PreconditionerSmoother::PointJacobi;
+  param.multigrid_data_pressure_poisson.coarse_problem.solver =
+    MultigridCoarseGridSolver::Chebyshev;
+  param.multigrid_data_pressure_poisson.coarse_problem.preconditioner =
+    MultigridCoarseGridPreconditioner::PointJacobi;
+
+  // pressure Poisson equation
+  param.solver_pressure_poisson              = SolverPressurePoisson::CG;
+  param.solver_data_pressure_poisson         = SolverData(1000, ABS_TOL, REL_TOL, 100);
+  param.preconditioner_pressure_poisson      = PreconditionerPressurePoisson::Multigrid;
+  param.multigrid_data_pressure_poisson.type = MultigridType::cphMG;
+  param.multigrid_data_pressure_poisson.coarse_problem.solver = MultigridCoarseGridSolver::CG;
+  param.multigrid_data_pressure_poisson.coarse_problem.preconditioner =
+    MultigridCoarseGridPreconditioner::AMG;
+
+  // projection step
+  param.solver_projection         = SolverProjection::CG;
+  param.solver_data_projection    = SolverData(1000, ABS_TOL, REL_TOL);
+  param.preconditioner_projection = PreconditionerProjection::InverseMassMatrix;
+
+
+  // HIGH-ORDER DUAL SPLITTING SCHEME
+
+  // formulations
+  param.order_extrapolation_pressure_nbc =
+    param.order_time_integrator <= 2 ? param.order_time_integrator : 2;
+
+  // viscous step
+  param.solver_viscous         = SolverViscous::CG;
+  param.solver_data_viscous    = SolverData(1000, ABS_TOL, REL_TOL);
+  param.preconditioner_viscous = PreconditionerViscous::InverseMassMatrix;
+
+
+  // PRESSURE-CORRECTION SCHEME
+
+  // formulation
+  param.order_pressure_extrapolation = 1; // use 0 for non-incremental formulation
+  param.rotational_formulation       = true;
+
+  // momentum step
+
+  // Newton solver
+  param.newton_solver_data_momentum = Newton::SolverData(100, ABS_TOL, REL_TOL);
+
+  // linear solver
+  param.solver_momentum = SolverMomentum::GMRES;
+  if(param.treatment_of_convective_term == TreatmentOfConvectiveTerm::Implicit)
+    param.solver_data_momentum = SolverData(1e4, ABS_TOL_LINEAR, REL_TOL_LINEAR, 100);
+  else
+    param.solver_data_momentum = SolverData(1e4, ABS_TOL, REL_TOL, 100);
+
+  param.preconditioner_momentum = MomentumPreconditioner::InverseMassMatrix;
+
+  // COUPLED NAVIER-STOKES SOLVER
+  param.use_scaling_continuity = false;
+
+  // nonlinear solver (Newton solver)
+  param.newton_solver_data_coupled = Newton::SolverData(100, ABS_TOL, REL_TOL);
+
+  // linear solver
+  param.solver_coupled = SolverCoupled::GMRES;
+  if(param.treatment_of_convective_term == TreatmentOfConvectiveTerm::Implicit)
+    param.solver_data_coupled = SolverData(1e3, ABS_TOL_LINEAR, REL_TOL_LINEAR, 100);
+  else
+    param.solver_data_coupled = SolverData(1e3, ABS_TOL, REL_TOL, 100);
+
+  // preconditioning linear solver
+  param.preconditioner_coupled = PreconditionerCoupled::BlockTriangular;
+
+  // preconditioner velocity/momentum block
+  param.preconditioner_velocity_block = MomentumPreconditioner::InverseMassMatrix;
+
+  // preconditioner Schur-complement block
+  param.preconditioner_pressure_block      = SchurComplementPreconditioner::CahouetChabard;
+  param.multigrid_data_pressure_block.type = MultigridType::cphMG;
+  param.multigrid_data_pressure_block.coarse_problem.solver = MultigridCoarseGridSolver::CG;
+  param.multigrid_data_pressure_block.coarse_problem.preconditioner =
+    MultigridCoarseGridPreconditioner::AMG;
+}
+
+namespace Precursor
 {
-public:
-  InflowProfile(InflowDataStorage<dim> const & inflow_data_storage)
-    : dealii::Function<dim>(dim, 0.0), data(inflow_data_storage)
-  {
-  }
-
-  double
-  value(dealii::Point<dim> const & p, unsigned int const component = 0) const final
-  {
-    double result = linear_interpolation_2d_cartesian(
-      p, data.y_values, data.z_values, data.velocity_values, component);
-
-    return result;
-  }
-
-private:
-  InflowDataStorage<dim> const & data;
-};
-
-template<int dim>
-class RightHandSide : public dealii::Function<dim>
-{
-public:
-  RightHandSide() : dealii::Function<dim>(dim, 0.0)
-  {
-  }
-
-  double
-  value(dealii::Point<dim> const & /*p*/, unsigned int const component = 0) const final
-  {
-    double result = 0.0;
-
-    // channel flow with periodic boundary conditions.
-    // The force is known (so no flow rate controller is used).
-    if(component == 0)
-      return 0.2844518;
-    else
-      return 0.0;
-
-    return result;
-  }
-};
-
-
 template<int dim, typename Number>
-class Application : public Precursor::ApplicationBase<dim, Number>
+class PrecursorDomain : public Domain<dim, Number>
 {
 public:
-  Application(std::string input_file, MPI_Comm const & comm)
-    : Precursor::ApplicationBase<dim, Number>(input_file, comm)
+  PrecursorDomain(std::string                               parameter_file,
+                  MPI_Comm const &                          comm,
+                  std::shared_ptr<InflowDataStorage<dim>> & inflow_data)
+    : Domain<dim, Number>(parameter_file, comm), inflow_data_storage(inflow_data)
   {
-    inflow_data_storage.reset(new InflowDataStorage<dim>(n_points_inflow));
-  }
-
-private:
-  /*
-   *  Most of the parameters are the same for both domains, so we write
-   *  this function for the actual domain and only "correct" the parameters
-   *  for the precursor by passing an additional parameter is_precursor.
-   */
-  void
-  do_set_parameters(Parameters & param, bool const is_precursor = false)
-  {
-    // MATHEMATICAL MODEL
-    param.problem_type                   = ProblemType::Unsteady;
-    param.equation_type                  = EquationType::NavierStokes;
-    param.use_outflow_bc_convective_term = true;
-    param.formulation_viscous_term       = FormulationViscousTerm::LaplaceFormulation;
-    param.formulation_convective_term    = FormulationConvectiveTerm::DivergenceFormulation;
-    param.right_hand_side                = true;
-
-
-    // PHYSICAL QUANTITIES
-    if(is_precursor)
-      param.start_time = precursor_start_time;
-    else
-      param.start_time = start_time;
-    param.end_time  = end_time;
-    param.viscosity = viscosity;
-
-
-    // TEMPORAL DISCRETIZATION
-    param.solver_type                     = SolverType::Unsteady;
-    param.temporal_discretization         = TemporalDiscretization::BDFDualSplittingScheme;
-    param.treatment_of_convective_term    = TreatmentOfConvectiveTerm::Explicit;
-    param.calculation_of_time_step_size   = TimeStepCalculation::CFL;
-    param.order_time_integrator           = 2;
-    param.start_with_low_order            = true;
-    param.adaptive_time_stepping          = true;
-    param.max_velocity                    = centerline_velocity;
-    param.cfl                             = 0.3;
-    param.cfl_exponent_fe_degree_velocity = 1.5;
-    param.time_step_size                  = 1.0e-1;
-    param.order_time_integrator           = 2;
-
-    // output of solver information
-    param.solver_info_data.interval_time = (end_time - start_time) / 100;
-
-    // SPATIAL DISCRETIZATION
-    param.grid.triangulation_type = TriangulationType::Distributed;
-    if(is_precursor)
-      param.grid.n_refine_global = param.grid.n_refine_global + additional_refinements_precursor;
-    param.mapping_degree = param.degree_u;
-    param.degree_p       = DegreePressure::MixedOrder;
-
-    // convective term
-    if(param.formulation_convective_term == FormulationConvectiveTerm::DivergenceFormulation)
-      param.upwind_factor = 0.5; // allows using larger CFL values for explicit formulations
-
-    // variant Direct allows to use larger time step
-    // sizes due to CFL condition at inflow boundary
-    param.type_dirichlet_bc_convective = TypeDirichletBCs::Mirror;
-
-    // viscous term
-    param.IP_formulation_viscous = InteriorPenaltyFormulation::SIPG;
-
-    // velocity pressure coupling terms
-    param.gradp_formulation = FormulationPressureGradientTerm::Weak;
-    param.divu_formulation  = FormulationVelocityDivergenceTerm::Weak;
-
-    // div-div and continuity penalty
-    param.use_divergence_penalty                     = true;
-    param.divergence_penalty_factor                  = 1.0e0;
-    param.use_continuity_penalty                     = true;
-    param.continuity_penalty_factor                  = param.divergence_penalty_factor;
-    param.continuity_penalty_components              = ContinuityPenaltyComponents::Normal;
-    param.apply_penalty_terms_in_postprocessing_step = true;
-    param.continuity_penalty_use_boundary_data       = true;
-
-    // TURBULENCE
-    param.turbulence_model_data.is_active        = false;
-    param.turbulence_model_data.turbulence_model = TurbulenceEddyViscosityModel::Sigma;
-    // Smagorinsky: 0.165
-    // Vreman: 0.28
-    // WALE: 0.50
-    // Sigma: 1.35
-    param.turbulence_model_data.constant = 1.35;
-
-    // PROJECTION METHODS
-
-    // pressure Poisson equation
-    param.solver_pressure_poisson              = SolverPressurePoisson::CG;
-    param.solver_data_pressure_poisson         = SolverData(1e4, 1.e-12, 1.e-6, 100);
-    param.preconditioner_pressure_poisson      = PreconditionerPressurePoisson::Multigrid;
-    param.multigrid_data_pressure_poisson.type = MultigridType::cphMG;
-    param.multigrid_data_pressure_poisson.smoother_data.smoother = MultigridSmoother::Chebyshev;
-    param.multigrid_data_pressure_poisson.smoother_data.preconditioner =
-      PreconditionerSmoother::PointJacobi;
-    param.multigrid_data_pressure_poisson.coarse_problem.solver =
-      MultigridCoarseGridSolver::Chebyshev;
-    param.multigrid_data_pressure_poisson.coarse_problem.preconditioner =
-      MultigridCoarseGridPreconditioner::PointJacobi;
-
-    // pressure Poisson equation
-    param.solver_pressure_poisson              = SolverPressurePoisson::CG;
-    param.solver_data_pressure_poisson         = SolverData(1000, ABS_TOL, REL_TOL, 100);
-    param.preconditioner_pressure_poisson      = PreconditionerPressurePoisson::Multigrid;
-    param.multigrid_data_pressure_poisson.type = MultigridType::cphMG;
-    param.multigrid_data_pressure_poisson.coarse_problem.solver = MultigridCoarseGridSolver::CG;
-    param.multigrid_data_pressure_poisson.coarse_problem.preconditioner =
-      MultigridCoarseGridPreconditioner::AMG;
-
-    // projection step
-    param.solver_projection         = SolverProjection::CG;
-    param.solver_data_projection    = SolverData(1000, ABS_TOL, REL_TOL);
-    param.preconditioner_projection = PreconditionerProjection::InverseMassMatrix;
-
-
-    // HIGH-ORDER DUAL SPLITTING SCHEME
-
-    // formulations
-    param.order_extrapolation_pressure_nbc =
-      param.order_time_integrator <= 2 ? param.order_time_integrator : 2;
-
-    // viscous step
-    param.solver_viscous         = SolverViscous::CG;
-    param.solver_data_viscous    = SolverData(1000, ABS_TOL, REL_TOL);
-    param.preconditioner_viscous = PreconditionerViscous::InverseMassMatrix;
-
-
-    // PRESSURE-CORRECTION SCHEME
-
-    // formulation
-    param.order_pressure_extrapolation = 1; // use 0 for non-incremental formulation
-    param.rotational_formulation       = true;
-
-    // momentum step
-
-    // Newton solver
-    param.newton_solver_data_momentum = Newton::SolverData(100, ABS_TOL, REL_TOL);
-
-    // linear solver
-    param.solver_momentum = SolverMomentum::GMRES;
-    if(param.treatment_of_convective_term == TreatmentOfConvectiveTerm::Implicit)
-      param.solver_data_momentum = SolverData(1e4, ABS_TOL_LINEAR, REL_TOL_LINEAR, 100);
-    else
-      param.solver_data_momentum = SolverData(1e4, ABS_TOL, REL_TOL, 100);
-
-    param.preconditioner_momentum = MomentumPreconditioner::InverseMassMatrix;
-
-    // COUPLED NAVIER-STOKES SOLVER
-    param.use_scaling_continuity = false;
-
-    // nonlinear solver (Newton solver)
-    param.newton_solver_data_coupled = Newton::SolverData(100, ABS_TOL, REL_TOL);
-
-    // linear solver
-    param.solver_coupled = SolverCoupled::GMRES;
-    if(param.treatment_of_convective_term == TreatmentOfConvectiveTerm::Implicit)
-      param.solver_data_coupled = SolverData(1e3, ABS_TOL_LINEAR, REL_TOL_LINEAR, 100);
-    else
-      param.solver_data_coupled = SolverData(1e3, ABS_TOL, REL_TOL, 100);
-
-    // preconditioning linear solver
-    param.preconditioner_coupled = PreconditionerCoupled::BlockTriangular;
-
-    // preconditioner velocity/momentum block
-    param.preconditioner_velocity_block = MomentumPreconditioner::InverseMassMatrix;
-
-    // preconditioner Schur-complement block
-    param.preconditioner_pressure_block      = SchurComplementPreconditioner::CahouetChabard;
-    param.multigrid_data_pressure_block.type = MultigridType::cphMG;
-    param.multigrid_data_pressure_block.coarse_problem.solver = MultigridCoarseGridSolver::CG;
-    param.multigrid_data_pressure_block.coarse_problem.preconditioner =
-      MultigridCoarseGridPreconditioner::AMG;
   }
 
   void
   set_parameters() final
   {
-    do_set_parameters(this->param);
+    this->param.degree_u             = this->resolution_parameters.degree;
+    this->param.grid.n_refine_global = this->resolution_parameters.refine_space;
+
+    do_set_parameters(this->param, true);
   }
 
   void
-  set_parameters_precursor() final
+  create_grid() final
   {
-    do_set_parameters(this->param_pre, true);
+    auto const lambda_create_triangulation =
+      [&](dealii::Triangulation<dim, dim> &                        tria,
+          std::vector<dealii::GridTools::PeriodicFacePair<
+            typename dealii::Triangulation<dim>::cell_iterator>> & periodic_face_pairs,
+          unsigned int const                                       global_refinements,
+          std::vector<unsigned int> const &                        vector_local_refinements) {
+        (void)periodic_face_pairs;
+        (void)vector_local_refinements;
+
+        Geometry::create_grid_precursor(tria, global_refinements, periodic_face_pairs);
+      };
+
+    GridUtilities::create_fine_and_coarse_triangulations<dim>(*this->grid,
+                                                              this->mpi_comm,
+                                                              this->param.grid,
+                                                              this->param.involves_h_multigrid(),
+                                                              lambda_create_triangulation,
+                                                              {} /* no local refinements */);
+  }
+
+  void
+  set_boundary_descriptor() final
+  {
+    typedef typename std::pair<dealii::types::boundary_id, std::shared_ptr<dealii::Function<dim>>>
+      pair;
+
+    // fill boundary descriptor velocity
+    // no slip boundaries at lower and upper wall with ID=0
+    this->boundary_descriptor->velocity->dirichlet_bc.insert(
+      pair(0, new dealii::Functions::ZeroFunction<dim>(dim)));
+
+    // fill boundary descriptor pressure
+    // no slip boundaries at lower and upper wall with ID=0
+    this->boundary_descriptor->pressure->neumann_bc.insert(0);
+  }
+
+  void
+  set_field_functions() final
+  {
+    this->field_functions->initial_solution_velocity.reset(
+      new InitialSolutionVelocity<dim>(centerline_velocity,
+                                       Geometry::LENGTH_CHANNEL,
+                                       Geometry::HEIGHT_CHANNEL,
+                                       Geometry::WIDTH_CHANNEL));
+    this->field_functions->initial_solution_pressure.reset(
+      new dealii::Functions::ZeroFunction<dim>(1));
+    this->field_functions->analytical_solution_pressure.reset(
+      new dealii::Functions::ZeroFunction<dim>(1));
+    this->field_functions->right_hand_side.reset(new RightHandSide<dim>());
+  }
+
+  std::shared_ptr<PostProcessorBase<dim, Number>>
+  create_postprocessor() final
+  {
+    std::shared_ptr<PostProcessorBase<dim, Number>> pp;
+
+    PostProcessorData<dim> pp_data;
+    // write output for visualization of results
+    pp_data.output_data.time_control_data.is_active        = this->output_parameters.write;
+    pp_data.output_data.time_control_data.start_time       = start_time;
+    pp_data.output_data.time_control_data.trigger_interval = (end_time - start_time) / 60.0;
+    pp_data.output_data.directory          = this->output_parameters.directory + "vtu/";
+    pp_data.output_data.filename           = this->output_parameters.filename + "_precursor";
+    pp_data.output_data.write_divergence   = true;
+    pp_data.output_data.write_q_criterion  = true;
+    pp_data.output_data.write_boundary_IDs = true;
+    pp_data.output_data.write_processor_id = true;
+    pp_data.output_data.degree             = this->param.degree_u;
+    pp_data.output_data.write_higher_order = false;
+
+    PostProcessorDataBFS<dim> pp_data_bfs;
+    pp_data_bfs.pp_data = pp_data;
+
+    // turbulent channel statistics
+    pp_data_bfs.turb_ch_data.time_control_data_statistics.time_control_data.is_active = true;
+    pp_data_bfs.turb_ch_data.time_control_data_statistics.time_control_data.start_time =
+      sample_start_time;
+    pp_data_bfs.turb_ch_data.time_control_data_statistics.time_control_data.end_time = end_time;
+    pp_data_bfs.turb_ch_data.time_control_data_statistics.time_control_data
+      .trigger_every_time_steps = sample_every_timesteps;
+    pp_data_bfs.turb_ch_data.time_control_data_statistics
+      .write_preliminary_results_every_nth_time_step = sample_every_timesteps * 100;
+
+
+    pp_data_bfs.turb_ch_data.cells_are_stretched = Geometry::use_grid_stretching_in_y_direction;
+    pp_data_bfs.turb_ch_data.viscosity           = viscosity;
+    pp_data_bfs.turb_ch_data.directory           = this->output_parameters.directory;
+    pp_data_bfs.turb_ch_data.filename            = this->output_parameters.filename + "_precursor";
+
+    // use turbulent channel data to prescribe inflow velocity for BFS
+    pp_data_bfs.inflow_data.write_inflow_data = true;
+    pp_data_bfs.inflow_data.normal_direction  = 0; /* x-direction */
+    pp_data_bfs.inflow_data.normal_coordinate = Geometry::X1_COORDINATE_OUTFLOW_CHANNEL;
+    pp_data_bfs.inflow_data.n_points_y        = inflow_data_storage->n_points_y;
+    pp_data_bfs.inflow_data.n_points_z        = inflow_data_storage->n_points_z;
+    pp_data_bfs.inflow_data.y_values          = &inflow_data_storage->y_values;
+    pp_data_bfs.inflow_data.z_values          = &inflow_data_storage->z_values;
+    pp_data_bfs.inflow_data.array             = &inflow_data_storage->velocity_values;
+
+    pp.reset(new PostProcessorBFS<dim, Number>(pp_data_bfs, this->mpi_comm));
+
+    return pp;
+  }
+
+private:
+  std::shared_ptr<InflowDataStorage<dim>> inflow_data_storage;
+};
+
+template<int dim, typename Number>
+class MainDomain : public Precursor::Domain<dim, Number>
+{
+public:
+  MainDomain(std::string                               parameter_file,
+             MPI_Comm const &                          comm,
+             std::shared_ptr<InflowDataStorage<dim>> & inflow_data)
+    : Domain<dim, Number>(parameter_file, comm), inflow_data_storage(inflow_data)
+  {
+  }
+
+  void
+  set_parameters() final
+  {
+    this->param.degree_u             = this->resolution_parameters.degree;
+    this->param.grid.n_refine_global = this->resolution_parameters.refine_space;
+
+    do_set_parameters(this->param, false);
   }
 
   void
@@ -359,30 +407,6 @@ private:
                                                               this->param.involves_h_multigrid(),
                                                               lambda_create_triangulation,
                                                               {} /* no local refinements */);
-  }
-
-  void
-  create_grid_precursor() final
-  {
-    auto const lambda_create_triangulation =
-      [&](dealii::Triangulation<dim, dim> &                        tria,
-          std::vector<dealii::GridTools::PeriodicFacePair<
-            typename dealii::Triangulation<dim>::cell_iterator>> & periodic_face_pairs,
-          unsigned int const                                       global_refinements,
-          std::vector<unsigned int> const &                        vector_local_refinements) {
-        (void)periodic_face_pairs;
-        (void)vector_local_refinements;
-
-        Geometry::create_grid_precursor(tria, global_refinements, periodic_face_pairs);
-      };
-
-    GridUtilities::create_fine_and_coarse_triangulations<dim>(
-      *this->grid_pre,
-      this->mpi_comm,
-      this->param_pre.grid,
-      this->param_pre.involves_h_multigrid(),
-      lambda_create_triangulation,
-      {} /* no local refinements */);
   }
 
   void
@@ -420,22 +444,6 @@ private:
   }
 
   void
-  set_boundary_descriptor_precursor() final
-  {
-    typedef typename std::pair<dealii::types::boundary_id, std::shared_ptr<dealii::Function<dim>>>
-      pair;
-
-    // fill boundary descriptor velocity
-    // no slip boundaries at lower and upper wall with ID=0
-    this->boundary_descriptor_pre->velocity->dirichlet_bc.insert(
-      pair(0, new dealii::Functions::ZeroFunction<dim>(dim)));
-
-    // fill boundary descriptor pressure
-    // no slip boundaries at lower and upper wall with ID=0
-    this->boundary_descriptor_pre->pressure->neumann_bc.insert(0);
-  }
-
-  void
   set_field_functions() final
   {
     this->field_functions->initial_solution_velocity.reset(
@@ -443,28 +451,13 @@ private:
                                        Geometry::LENGTH_CHANNEL,
                                        Geometry::HEIGHT_CHANNEL,
                                        Geometry::WIDTH_CHANNEL));
-    //  this->field_functions->initial_solution_velocity.reset(new
-    //  dealii::Functions::ZeroFunction<dim>(dim));
+    this->field_functions->initial_solution_velocity.reset(
+      new dealii::Functions::ZeroFunction<dim>(dim));
     this->field_functions->initial_solution_pressure.reset(
       new dealii::Functions::ZeroFunction<dim>(1));
     this->field_functions->analytical_solution_pressure.reset(
       new dealii::Functions::ZeroFunction<dim>(1));
     this->field_functions->right_hand_side.reset(new dealii::Functions::ZeroFunction<dim>(dim));
-  }
-
-  void
-  set_field_functions_precursor() final
-  {
-    this->field_functions_pre->initial_solution_velocity.reset(
-      new InitialSolutionVelocity<dim>(centerline_velocity,
-                                       Geometry::LENGTH_CHANNEL,
-                                       Geometry::HEIGHT_CHANNEL,
-                                       Geometry::WIDTH_CHANNEL));
-    this->field_functions_pre->initial_solution_pressure.reset(
-      new dealii::Functions::ZeroFunction<dim>(1));
-    this->field_functions_pre->analytical_solution_pressure.reset(
-      new dealii::Functions::ZeroFunction<dim>(1));
-    this->field_functions_pre->right_hand_side.reset(new RightHandSide<dim>());
   }
 
   std::shared_ptr<PostProcessorBase<dim, Number>>
@@ -729,98 +722,33 @@ private:
     return pp;
   }
 
-  std::shared_ptr<PostProcessorBase<dim, Number>>
-  create_postprocessor_precursor() final
-  {
-    std::shared_ptr<PostProcessorBase<dim, Number>> pp;
-
-    PostProcessorData<dim> pp_data;
-    // write output for visualization of results
-    pp_data.output_data.time_control_data.is_active        = this->output_parameters.write;
-    pp_data.output_data.time_control_data.start_time       = start_time;
-    pp_data.output_data.time_control_data.trigger_interval = (end_time - start_time) / 60.0;
-    pp_data.output_data.directory          = this->output_parameters.directory + "vtu/";
-    pp_data.output_data.filename           = this->output_parameters.filename + "_precursor";
-    pp_data.output_data.write_divergence   = true;
-    pp_data.output_data.write_q_criterion  = true;
-    pp_data.output_data.write_boundary_IDs = true;
-    pp_data.output_data.write_processor_id = true;
-    pp_data.output_data.degree             = this->param_pre.degree_u;
-    pp_data.output_data.write_higher_order = false;
-
-    PostProcessorDataBFS<dim> pp_data_bfs;
-    pp_data_bfs.pp_data = pp_data;
-
-    // turbulent channel statistics
-    pp_data_bfs.turb_ch_data.time_control_data_statistics.time_control_data.is_active = true;
-    pp_data_bfs.turb_ch_data.time_control_data_statistics.time_control_data.start_time =
-      sample_start_time;
-    pp_data_bfs.turb_ch_data.time_control_data_statistics.time_control_data.end_time = end_time;
-    pp_data_bfs.turb_ch_data.time_control_data_statistics.time_control_data
-      .trigger_every_time_steps = sample_every_timesteps;
-    pp_data_bfs.turb_ch_data.time_control_data_statistics
-      .write_preliminary_results_every_nth_time_step = sample_every_timesteps * 100;
-
-
-    pp_data_bfs.turb_ch_data.cells_are_stretched = Geometry::use_grid_stretching_in_y_direction;
-    pp_data_bfs.turb_ch_data.viscosity           = viscosity;
-    pp_data_bfs.turb_ch_data.directory           = this->output_parameters.directory;
-    pp_data_bfs.turb_ch_data.filename            = this->output_parameters.filename + "_precursor";
-
-    // use turbulent channel data to prescribe inflow velocity for BFS
-    pp_data_bfs.inflow_data.write_inflow_data = true;
-    pp_data_bfs.inflow_data.normal_direction  = 0; /* x-direction */
-    pp_data_bfs.inflow_data.normal_coordinate = Geometry::X1_COORDINATE_OUTFLOW_CHANNEL;
-    pp_data_bfs.inflow_data.n_points_y        = inflow_data_storage->n_points_y;
-    pp_data_bfs.inflow_data.n_points_z        = inflow_data_storage->n_points_z;
-    pp_data_bfs.inflow_data.y_values          = &inflow_data_storage->y_values;
-    pp_data_bfs.inflow_data.z_values          = &inflow_data_storage->z_values;
-    pp_data_bfs.inflow_data.array             = &inflow_data_storage->velocity_values;
-
-    pp.reset(new PostProcessorBFS<dim, Number>(pp_data_bfs, this->mpi_comm));
-
-    return pp;
-  }
-
-  // consider a friction Reynolds number of Re_tau = u_tau * H / nu = 290
-  // and body force f = tau_w/H with tau_w = u_tau^2.
-  double const viscosity = 1.5268e-5;
-
-  // precursor simulation: data structures for storage of inflow data
+private:
   std::shared_ptr<InflowDataStorage<dim>> inflow_data_storage;
-
-  // number of points for inflow boundary condition
-  unsigned int const n_points_inflow = 101;
-
-  unsigned int const additional_refinements_precursor = 0;
-
-  // start and end time
-  double const Re_H                 = 5540.0;
-  double const centerline_velocity  = Re_H * viscosity / Geometry::H;
-  double const characteristic_time  = Geometry::H / centerline_velocity;
-  double const start_time           = 0.0;
-  double const precursor_start_time = -300.0 * characteristic_time;
-  double const end_time             = 300.0 * characteristic_time;
-
-  // postprocessing
-
-  // sampling of statistical results
-  double const       sample_start_time      = 100.0 * characteristic_time;
-  unsigned int const sample_every_timesteps = 10;
-  unsigned int const n_points_per_line      = 101;
-
-  // solver tolerances
-  double const ABS_TOL = 1.e-12;
-  double const REL_TOL = 1.e-3;
-
-  double const ABS_TOL_LINEAR = 1.e-12;
-  double const REL_TOL_LINEAR = 1.e-2;
 };
 
-} // namespace IncNS
+template<int dim, typename Number>
+class Application : public Precursor::ApplicationBase<dim, Number>
+{
+public:
+  Application(std::string input_file, MPI_Comm const & comm)
+    : ApplicationBase<dim, Number>(input_file, comm)
+  {
+    inflow_data_storage.reset(new InflowDataStorage<dim>(n_points_inflow));
 
+    this->precursor =
+      std::make_shared<PrecursorDomain<dim, Number>>(input_file, comm, inflow_data_storage);
+    this->main = std::make_shared<MainDomain<dim, Number>>(input_file, comm, inflow_data_storage);
+  }
+
+private:
+  // precursor simulation: data structures for storage of inflow data
+  std::shared_ptr<InflowDataStorage<dim>> inflow_data_storage;
+};
+
+} // namespace Precursor
+} // namespace IncNS
 } // namespace ExaDG
 
 #include <exadg/incompressible_navier_stokes/user_interface/implement_get_application_precursor.h>
 
-#endif /* APPLICATIONS_INCOMPRESSIBLE_NAVIER_STOKES_TEST_CASES_TURBULENT_CHANNEL_H_ */
+#endif /* APPLICATIONS_INCOMPRESSIBLE_NAVIER_STOKES_TEST_CASES_BFS_H_ */
